Improved code design by making a class that contains our workbook. Same functionality as before, different class names
